api/requirements.txt
alembic==1.12.0
bcrypt==4.0.1
black==23.7.0
cryptography==41.0.3
Flask==2.3.3
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.5
google-auth==2.22.0
icecream==2.1.3
logtail-python==0.2.6
mysqlclient==2.2.0
pendulum==2.1.2
pydantic==2.3.0
PyJWT==2.8.0
pylint==2.17.5
pymysql==1.1.0
pytest==7.4.1
python-dotenv==1.0.0
requests==2.31.0
sendgrid==6.10.0
sentry-sdk==1.31.0
stripe==6.4.0